requirements.txt
pcapy
scapy-python3
pyfuzz==0.1.3
Pillow